# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API MathEasy",
        "description": "API per gestionar operacions i recursos de MathEasy amb Laravel i autenticació Bearer",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://192.168.1.39:8000/api",
            "description": "Servidor local de desenvolupament"
        }
    ],
    "paths": {
...
        "/informes": {
            "get": {
                "tags": [
                    "Informes"
                ],
                "summary": "Llistar tots els informes",
                "description": "Retorna una llista amb tots els informes creats",
                "operationId": "427188f4b0e442f4d4a2bf1f2ff7c66f",
                "responses": {
                    "200": {
                        "description": "Llista d'informes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Informe"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Usuari no autenticat"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Informes"
                ],
                "summary": "Crear un nou informe",
                "description": "Crea un informe amb les dades proporcionades",
                "operationId": "5913480bd8bc3db0d68df030c3ea9bb2",
                "requestBody": {
                    "description": "Dades de l'informe",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InformeInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Informe creat correctament",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Informe"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dades no vàlides"
                    },
                    "401": {
                        "description": "Usuari no autenticat"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/informesAlumne/{alumne}": {
            "get": {
                "tags": [
                    "Informes"
                ],
                "summary": "Llistar tots els informes d'un alumne concret",
                "description": "Retorna una llista amb tots els informes creats d'un alumne concret",
                "operationId": "32de57bfe360d0c94eca3faf569c6683",
                "parameters": [
                    {
                        "name": "alumne",
                        "in": "path",
                        "description": "ID de l'alumne",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Llista d'informes d'un alumne concret",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Informe"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Alumne no trobat"
                    },
                    "401": {
                        "description": "Usuari no autenticat"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Informe": {
                "title": "Informe",
                "description": "Representa un informe",
                "properties": {
                    "id": {
                        "description": "ID de l'informe",
                        "type": "integer"
                    },
                    "Tipus_partida": {
                        "description": "Tipus de partida",
                        "type": "string"
                    },
                    "Respostes_correctes": {
                        "description": "Numero de respostes correctes",
                        "type": "integer"
                    },
                    "Respostes_incorrectes": {
                        "description": "Numero de respostes incorrectes",
                        "type": "integer"
                    },
                    "Experiencia": {
                        "description": "Experiencia obtinguda",
                        "type": "integer"
                    },
                    "alumne_id": {
                        "description": "ID de l'alumne",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "InformeInput": {
                "title": "Entrada Informe",
                "description": "Dades necessàries per crear un informe",
                "properties": {
                    "Tipus_partida": {
                        "description": "Tipus de partida",
                        "type": "string"
                    },
                    "Respostes_correctes": {
                        "description": "Numero de respostes correctes",
                        "type": "integer"
                    },
                    "Respostes_incorrectes": {
                        "description": "Numero de respostes incorrectes",
                        "type": "integer"
                    },
                    "Experiencia": {
                        "description": "Experiencia obtinguda",
                        "type": "integer"
                    },
                    "alumne_id": {
                        "description": "ID de l'alumne",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Introduïu el token Bearer JWT per autenticar-vos",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}